Add alerts state
Needed an aggregation method, to not fetch the first occurrence but the most restrictive one.
What does this mean? If there are 3 alert rules, and one is firing, the overall state should be firing as well,
even if the other 2 are normal.

This is based on the work done here: https://github.com/K-Phoen/backstage-plugin-grafana/pull/54

Also, added the state in the table now, instead of hidding it if the unified alerting is used.

The method is now allowing an array of strings or a single string. This simplifies the whole call process, which should be
much faster than making a call per selector foundas before